+++ coffang/src/main/java/com/sparta/coffang/controller/CoffeeController.java
@@ -3,10 +3,8 @@ package com.sparta.coffang.controller;
 
 import com.sparta.coffang.dto.PhotoDto;
 import com.sparta.coffang.dto.requestDto.CoffeeRequestDto;
-import com.sparta.coffang.dto.responseDto.CoffeeResponseDto;
 import com.sparta.coffang.exceptionHandler.CustomException;
 import com.sparta.coffang.exceptionHandler.ErrorCode;
-import com.sparta.coffang.model.Coffee;
 import com.sparta.coffang.model.UserRoleEnum;
 import com.sparta.coffang.security.UserDetailsImpl;
 import com.sparta.coffang.service.CoffeeService;
@@ -17,7 +15,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import javax.transaction.Transactional;
 import java.util.List;
 
 @RestController
@@ -91,6 +88,21 @@ public class CoffeeController {
             return coffeeService.getByCategory(category);
 
         throw new CustomException(ErrorCode.API_NOT_FOUND);
+    }
+
+
+    //커피 이미지만 1개 등록
+    @PostMapping("/coffee/image")
+    public ResponseEntity imageUpload(@RequestPart("imgUrl") List<MultipartFile> multipartFiles,
+                                      @AuthenticationPrincipal UserDetailsImpl userDetails) {
+
+        List<PhotoDto> photoDtos = s3Service.uploadFile(multipartFiles);
+        return coffeeService.imageUpload(photoDtos.get(0));
+    }
+    //커피 이미지만 1개 조회
+    @GetMapping("/coffee/image/{imageId}")
+    public ResponseEntity getImage(@PathVariable Long imageId) {
+        return coffeeService.getImage(imageId);
     }
 }
