+++ coffang/src/main/java/com/sparta/coffang/dto/responseDto/ImageResponseDto.java
@@ -1,0 +1,19 @@
+package com.sparta.coffang.dto.responseDto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ImageResponseDto {
+
+    private Long imageId;
+
+    private String img;
+}
